requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
groq
tiktoken
python-dotenv
